+++ ft_put_ptr.c
@@ -6,13 +6,13 @@
 /*   By: stena-he <stena-he@student.42wolfsburg.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/03 18:33:07 by stena-he          #+#    #+#             */
-/*   Updated: 2022/07/03 18:57:46 by stena-he         ###   ########.fr       */
+/*   Updated: 2022/07/03 19:51:28 by stena-he         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void	ft_put_pointer(unsigned long long ptr)
+void	ft_put_pointer(unsigned long ptr)
 {
 	write(1, "0x", 2);
 	if (ptr == 0)
